vspmfilter: Rewrite implement for swap
Old implement decide the swap base on input and out VSP format.
This is not good:
- Troublesome to add new format
- Can not add many GStreamer formats which use same VSP format
(but difference in swap)

This patch fixes it.
Besides, also fix mistakes in calling function set_colorspace
(use same set_colorspace for both input and output in tranformm_frame,
use output format with set_colorspace for input in transform_buffer)

Signed-off-by: HungTran <hung.tran.jy@rvc.renesas.com>