vspmfilter:Fix crash issue in dmabuf-use mode with some resolutions
In some output resolution such as 960x540 or 320x240, dmabuf-use mode
causes kernel crash.

This is because in dmabuf-use mode, waylandsink will map dmabuf created
by this plugin, but it only allows memory areas that are page-aligned.
In resolutions that UY plane is not page-aligned (such as above),
mapping dmabuf will fail can cause kernel crash.

This patch fixes this issue by increase the plane size to ensure
page aligned in all resolutions.
Besides, it also updates the transform_buffer function to
convert correctly in the resolution that width is not
dividable for 32 (such as 720x480)

Signed-off-by: HungTran <hung.tran.jy@rvc.renesas.com>