vspmfilter: Fix wrong stride numbers
Previously vspmfilter guesses the stride numbers from video foramt.
But this is not a correct method, upstream and downstream elements
can use different strides. When this happens, the output video
is not correct.
Especially when run with omxh264enc, the strides in omxh264enc is
largely different with video info due to its alignment requirement.

This wrong stride issue is fix here, use the stride info inside
frame info (owner of buffer must set stride data correctly in the
video meta at first)

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>