vspmfilter: enable passthrough
This commit enables passthrough mode.
After enabled, vspmfilter will passthrough the buffer without
processing if input caps and output caps are same.
This improve performance in that case.

Besides, this fixes the wrong output image issue happens when
2 instances of vspmfilter are used as a same time (which happens
a lot in playbin3).
As a result, playbin3 can run with good performance after
this commit with below command:
    gst-launch-1.0 playbin uri=file:///<video> video-sink=waylandsink

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>