Feature: Support ISU_RAW10
RAW10 is a variant of gray scale packed into 64bits little endian words.
The MSB 4 bits are padding should be ignored. This format is compatible with
GST_VIDEO_FORMAT_GRAY10_LE64 which is defined by Renesas.

The stride of this format must be calculated separately with the other common
formats because there is no clear definition of p_stride.

Signed-off-by: Kiet Pham <kiet.pham.xb@renesas.com>